Código organizado e tentativa de solução de problema com css
Código foi organizado de maneira diferente e problema onde css não aparece está em andamento para resolução, mas ainda não resolvido.
