+++ backend/src/home/home.html
@@ -1,0 +1,28 @@
+<!DOCTYPE html>
+<html lang="pt-BR">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Home</title>
+</head>
+<body>
+    <header>Home</header>
+    <br>
+    <main>
+        <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+            Donec mi nulla, tincidunt rhoncus consequat ut, accumsan in mauris.
+            Phasellus sit amet justo non nibh placerat convallis. 
+            Phasellus convallis nisl odio, in pellentesque quam consectetur ut. 
+            Maecenas eu semper nunc. Curabitur libero risus, imperdiet et nunc non, accumsan tempor ex. 
+            Nam velit dui, faucibus non metus non, auctor sodales tortor. 
+            Curabitur sagittis nunc lacus, imperdiet blandit orci aliquam eu. 
+            Sed finibus nunc magna. Ut ex velit, egestas sed urna elementum, semper accumsan sapien. 
+            Etiam viverra maximus elementum. Praesent id elementum urna. 
+            Donec at laoreet quam, eu aliquet diam. Suspendisse sed lacus quis dolor vehicula aliquet vel a ipsum. 
+            Vivamus vel lectus molestie, pretium tellus vitae, commodo odio. 
+            Aenean nec turpis ac nunc interdum sagittis et quis est. In id tortor efficitur, vestibulum dolor id, pharetra metus.
+        </p>
+    </main>
+    <footer>2023</footer>
+</body>
+</html>
