+++ backend/src/login/login.html
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html lang="pt-BR">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Login</title>
+</head>
+<body>
+    <h1>Login</h1>
+    <form id="login">
+        <label for="mail">Email:* </label>
+        <input type="email" id="mail_login" placeholder="Digite seu Email" required>
+        <br>
+        <label for="senha">Senha:* </label>
+        <input type="password" id="senha_login" placeholder="Digite sua senha" required>
+        <br>
+        <input type="submit" id="btn_envio" value="Enviar">
+    </form>
+</body>
+</html>
