--- sistema/src/cadastro/cadastro.html
@@ -1,29 +1,0 @@
-<!DOCTYPE html>
-<html lang="pt-BR">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Cadastro</title>
-</head>
-<body>
-    <h1>Cadastro</h1>
-    <form id="cadastro">
-        <label for="nome">Nome Completo:* </label>
-        <input type="text" id="nome" placeholder="Digite seu nome" required>
-        <br>
-        <label for="cpf">CPF:* </label>
-        <input type="number" id="cpf" placeholder="Digite seu CPF" required>
-        <br>
-        <label for="mail">Email:* </label>
-        <input type="email" id="mail_cadastro" placeholder="Digite seu Email" required>
-        <br>
-        <label for="telefone">Telefone:* </label>
-        <input type="number" id="telefone" placeholder="Digite seu nº de telefone" required>
-        <br>
-        <label for="senha">Senha:* </label>
-        <input type="password" id="senha_cadastro" placeholder="Digite uma senha" required>
-        <br>
-        <input type="submit" id="btn_envio" value="Enviar">
-    </form>
-</body>
-</html>
