--- sistema/src/contato/contato.html
@@ -1,28 +1,0 @@
-<!DOCTYPE html>
-<html lang="pt-BR">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Contato</title>
-</head>
-<body>
-    <header>Contato</header>
-    <br>
-    <main>
-        <h1>Fale Conosco!</h1>
-        <form id="contato">
-            <label for="nome_contato">Nome Completo: </label>
-            <input type="text" id="nome_contato" placeholder="Digite seu nome">
-            <br>
-            <label for="mail">Email:* </label>
-            <input type="email" id="mail_contato" placeholder="Digite seu Email" required>
-            <br>
-            <label for="descricao_prob">Descrição do Problema:*</label>
-            <input type="text" id="descricao_prob" placeholder="Informe seu problema" required>
-            <br>
-            <input type="submit" id="btn_envio" value="Enviar">
-        </form>
-    </main>
-    <footer>2023</footer>
-</body>
-</html>
